+++ docs/main/changelog.rst
@@ -7,6 +7,12 @@ that were made in every particular version.
 From version 0.7.6 *Dependency Injector* framework strictly 
 follows `Semantic versioning`_
 
+4.24.0
+------
+- Add docs on ``@containers.copy()`` decorator.
+- Refactor ``@containers.copy()`` decorator.
+- Refactor async mode support in containers module.
+
 4.23.5
 ------
 - Fix docs publishing.